Create README.md
This project is for a job application. Solita has given an awesome task to everyone who applies to the dev academy. I copied the task assignment here below:

The exercise
Create a web application that uses a backend service to fetch the data. Backend can be made with any technology. We at Solita use for example (not in preference order) Java/Kotlin/Clojure/C#/TypeScript/Go but you are free to choose any other technology as well.

Backend can use a database, or it can be memory-based. Real database use is a preferable choice because it allows you to show broader skills. Also, the datasets are quite big so in-memory operations may be quite slow.

You can also freely choose the frontend (and possibly mobile frontend) technologies to use. The important part is to give good instructions on how to build and run the project.


above, you can see a small excerpt from the task, if you want to view more detailed instructions, see the link:
https://github.com/solita/dev-academy-2023-exercise#the-exercise